Fix [PageController.page cannot be accessed before a PageView is built with it]
a PageController cannot be fiddled with until a PageView has been already built with it. Hence we are using a function `_afterBuild` which will be called back by `addPostFrameCallBack` which makes sure that it will only be called back once the first frame is displayed. i.e, the build has finished.